GH-483 Extract Exception Handler
